Rename 'justspacesapp' to 'surveys'
