+++ surveys/apps.py
@@ -1,0 +1,5 @@
+rom django.apps import AppConfig
+
+
+class SurveysConfig(AppConfig):
+    name = 'surveys'
