+++ surveys/templates/study_create.html
